validate allowed hostnames more correctly
Previously, a hostname in a DNS request was validated by checking that
it was a suffix of an allowed hostname. That worked fine, but introduced
a potential bypass: querying for a disallowed hostname that has an
allowed hostname as its suffix. For example, if "google.com" was allowed,
using the original logic "ggoogle.com" would be allowed as well.

The logic is updated to check if the hostname(s) in DNS requests and
responses are either an exact match of an allowed hostname, or a
subdomain of an allowed hostname, which is what I originally intended
anyway.